refactor(ai-blog-writer): split itinerary ai workflows
