+++ apps/ai-blog-writer/apps/frontend/src/features/listicleItineraries/builder/hooks/itineraryBuilderAiActions.types.ts
@@ -1,0 +1,26 @@
+import type { Dispatch, SetStateAction } from 'react'
+import type {
+  ItineraryBlockType,
+  ListicleItineraryDraft,
+  LocationOption,
+  RelatedItemOption
+} from '../../types'
+
+export type StopComposeChoice = {
+  itemId: string
+  dayIndex: number
+  targetId: string
+  stopTitle: string
+  strandsNeighbor: boolean
+}
+
+export type ItineraryBuilderAiActionsParams = {
+  token?: string | null
+  draft: ListicleItineraryDraft | null
+  setDraft: Dispatch<SetStateAction<ListicleItineraryDraft | null>>
+  locations: LocationOption[]
+  relatedByBlockType: Record<ItineraryBlockType, RelatedItemOption[]>
+  canonicalStructuredData: string
+  onError: (message: string) => void
+  setResult: Dispatch<SetStateAction<string | null>>
+}
